+++ PgRoutiner/Builder/CodeBuilder/CodeRoutinesBuilder.cs
@@ -13,17 +13,12 @@ namespace PgRoutiner
         {
         }
 
-        protected override Module GetModule(Settings settings, CodeSettings codeSettings) => new RoutineModule(settings, codeSettings);
-
-        protected override IEnumerable<(Code code, string name, string shortFilename, string fullFileName, string relative, Module module)> 
-            GetCodes(string outputDir)
+        protected override IEnumerable<CodeResult> GetCodes()
         {
             foreach (var group in connection.GetRoutineGroups(settings, all: false))
             {
-                var name = group.Key;
-                var (shortFilename, fullFileName, relative) = GetFileNames(name, outputDir);
-
-                var module = GetModule(settings, codeSettings);
+                var name = group.Key.Name;
+                var module = new RoutineModule(settings, codeSettings, group.Key.Schema);
                 Code code;
                 try
                 {
@@ -31,10 +26,16 @@ namespace PgRoutiner
                 }
                 catch (ArgumentException e)
                 {
-                    Builder.Error($"File {relative} could not be generated. {e.Message}");
+                    Builder.Error($"Code for routine {name} could not be generated. {e.Message}");
                     continue;
                 }
-                yield return (code, name, shortFilename, fullFileName, relative, module);
+                yield return new CodeResult
+                { 
+                    Code = code, 
+                    Name = name,  
+                    Module = module,
+                    Schema = group.Key.Schema
+                };
             }
         }
     }