+++ PgRoutiner/DataAccess/GetTables.cs
@@ -9,20 +9,19 @@ namespace PgRoutiner
 {
     public static partial class DataAccess
     {
-        public static IEnumerable<PgItem> GetTables(this NpgsqlConnection connection, Settings settings) => 
-            connection.Read<(string Schema, string Name, string Type)>(@$"
+        public static IEnumerable<PgItem> GetTables(this NpgsqlConnection connection, Settings settings)
+        {
+            return connection.Read<(string Schema, string Name, string Type)>(@$"
 
             select 
                 table_schema, table_name, table_type
             from 
                 information_schema.tables
             where
-                (   @schema is not null and table_schema similar to @schema   )
-                or
-                (   {GetSchemaExpression("table_schema")}  )
+                (   @schema is null or (table_schema similar to @schema)   )
+                and (   {GetSchemaExpression("table_schema")}  )
 
-            ", ("schema", settings.Schema, DbType.AnsiString))
-            .Select(t => new PgItem
+            ", ("schema", settings.Schema, DbType.AnsiString)).Select(t => new PgItem
             {
                 Schema = t.Schema,
                 Name = t.Name,
@@ -34,5 +33,6 @@ namespace PgRoutiner
                     _ => PgType.Unknown
                 }
             });
+        }
     }
 }